+++ app/src/main/res/layout/fragment_contact_list_activity.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ContactListFragmentActivity">
+    tools:context=".fragment.ContactListFragmentActivity">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
